Allow outbound-api messages to pass the sms1 obj.
If a number is part of a messaging service, its messages will show up as "Outgoing API" requests instead of outbound replies. I believe we want to allow this to pass this objective. I think the main intention of this check was just to make sure the message direction wasn't inbound.

More info about this in Discord, it was reported by a player:
https://discord.com/channels/570292611215130634/804374967714054144/886181579301457931

`outbound-api` is the direction for allowing outbound api requests: https://www.twilio.com/docs/sms/api/message-resource